+++ src/controllers/CacheController.php
@@ -148,7 +148,7 @@ class CacheController extends Controller
         Blitz::$plugin->refreshCache->refreshAll();
         $message = 'Blitz cache successfully refreshed.';
 
-        if (Blitz::$plugin->settings->shouldGenerateOnRefresh()) {
+        if (Blitz::$plugin->settings->generateOnRefresh()) {
             $message = 'Blitz cache successfully refreshed and queued for generation.';
         }
 
@@ -183,7 +183,7 @@ class CacheController extends Controller
         Blitz::$plugin->refreshCache->refreshSite($siteId);
         $message = 'Site successfully refreshed.';
 
-        if (Blitz::$plugin->settings->shouldGenerateOnRefresh()) {
+        if (Blitz::$plugin->settings->generateOnRefresh()) {
             $message = 'Site successfully refreshed and queued for generation.';
         }